moving proof related DS to proof_generation directory
