--- lib/identity/data/data_sources/local_files_data_source.dart
@@ -1,15 +1,0 @@
-/*
-import 'dart:typed_data';
-
-import 'package:flutter/services.dart';
-
-class LocalFilesDataSource {
-  Future<List<Uint8List>> loadCircuitFiles(String circuitId) async {
-    final circuitDatPath = 'assets/auth/$circuitId.dat';
-    final circuitProvingKeyPath = 'assets/auth/$circuitId.zkey';
-    ByteData datFile = await rootBundle.load(circuitDatPath);
-    ByteData zkeyFile = await rootBundle.load(circuitProvingKeyPath);
-    return [datFile.buffer.asUint8List(), zkeyFile.buffer.asUint8List()];
-  }
-}
-*/
