--- lib/identity/data/dtos/witness_type.dart
@@ -1,1 +1,0 @@
-enum WitnessType { def, mtp, sig }
