+++ lib/identity/domain/exceptions/identity_exceptions.dart
@@ -12,8 +12,4 @@ class UnknownIdentityException implements Exception {
   UnknownIdentityException(this.identifier);
 }
 
-class NullWitnessException implements Exception {
-  final String? circuit;
 
-  NullWitnessException(this.circuit);
-}
