+++ lib/identity/domain/use_cases/authenticate_use_case.dart
@@ -1,15 +1,12 @@
 import 'package:polygonid_flutter_sdk/common/domain/use_case.dart';
 import 'package:polygonid_flutter_sdk/identity/domain/repositories/identity_repository.dart';
-import 'package:polygonid_flutter_sdk/proof_generation/domain/entities/circuit_data_entity.dart';
 
 class AuthenticateParam {
   final String issuerMessage;
-  final CircuitDataEntity circuitDataEntity;
   final String identifier;
 
   AuthenticateParam({
     required this.issuerMessage,
-    required this.circuitDataEntity,
     required this.identifier,
   });
 }
@@ -24,7 +21,6 @@ class AuthenticateUseCase extends FutureUseCase<AuthenticateParam, void> {
     try {
       await _identityRepository.authenticate(
         issuerMessage: param.issuerMessage,
-        circuitDataEntity: param.circuitDataEntity,
         identifier: param.identifier,
       );
     } catch (_) {}
