+++ lib/proof_generation/data/dtos/prove_param.dart
@@ -1,0 +1,8 @@
+import 'dart:typed_data';
+
+class ProveParam {
+  final Uint8List zKey;
+  final Uint8List wtns;
+
+  ProveParam(this.zKey, this.wtns);
+}
