+++ lib/proof_generation/data/dtos/witness_param.dart
@@ -1,0 +1,11 @@
+import 'dart:typed_data';
+
+class WitnessParam {
+  final Uint8List wasm;
+  final Uint8List json;
+
+  WitnessParam({
+    required this.wasm,
+    required this.json,
+  });
+}
