+++ lib/proof_generation/data/dtos/witness_type.dart
@@ -1,0 +1,1 @@
+enum WitnessType { auth, mtp, sig }
