+++ lib/proof_generation/domain/exceptions/proof_generation_exceptions.dart
@@ -1,0 +1,11 @@
+import 'package:polygonid_flutter_sdk/common/domain/error_exception.dart';
+
+class ProofGenerationException extends ErrorException {
+  ProofGenerationException(error) : super(error);
+}
+
+class NullWitnessException implements Exception {
+  final String? circuit;
+
+  NullWitnessException(this.circuit);
+}
