+++ lib/sdk/di/injector.config.dart
@@ -10,61 +10,68 @@ import 'package:http/http.dart' as _i7;
 import 'package:injectable/injectable.dart' as _i2;
 import 'package:sembast/sembast.dart' as _i9;
 
-import '../../credential/data/credential_repository_impl.dart' as _i33;
+import '../../credential/data/credential_repository_impl.dart' as _i35;
 import '../../credential/data/data_sources/remote_claim_data_source.dart'
-    as _i22;
+    as _i23;
 import '../../credential/data/data_sources/storage_claim_data_source.dart'
-    as _i30;
-import '../../credential/data/mappers/claim_mapper.dart' as _i29;
+    as _i32;
+import '../../credential/data/mappers/claim_mapper.dart' as _i31;
 import '../../credential/data/mappers/claim_state_mapper.dart' as _i6;
 import '../../credential/data/mappers/credential_request_mapper.dart' as _i8;
 import '../../credential/data/mappers/filter_mapper.dart' as _i10;
 import '../../credential/data/mappers/filters_mapper.dart' as _i11;
 import '../../credential/data/mappers/id_filter_mapper.dart' as _i13;
 import '../../credential/domain/repositories/credential_repository.dart'
-    as _i34;
+    as _i36;
 import '../../credential/domain/use_cases/fetch_and_save_claims_use_case.dart'
-    as _i48;
-import '../../credential/domain/use_cases/get_claims_use_case.dart' as _i35;
-import '../../credential/domain/use_cases/remove_claims_use_case.dart' as _i37;
-import '../../credential/domain/use_cases/update_claim_use_case.dart' as _i38;
+    as _i50;
+import '../../credential/domain/use_cases/get_claims_use_case.dart' as _i37;
+import '../../credential/domain/use_cases/remove_claims_use_case.dart' as _i39;
+import '../../credential/domain/use_cases/update_claim_use_case.dart' as _i40;
 import '../../identity/data/data_sources/jwz_data_source.dart' as _i16;
 import '../../identity/data/data_sources/lib_identity_data_source.dart' as _i17;
-import '../../identity/data/data_sources/proof_scope_data_source.dart' as _i19;
-import '../../identity/data/data_sources/prover_lib_data_source.dart' as _i21;
 import '../../identity/data/data_sources/remote_identity_data_source.dart'
-    as _i23;
+    as _i24;
 import '../../identity/data/data_sources/storage_identity_data_source.dart'
-    as _i31;
+    as _i33;
 import '../../identity/data/data_sources/storage_key_value_data_source.dart'
-    as _i32;
-import '../../identity/data/data_sources/wallet_data_source.dart' as _i24;
-import '../../identity/data/data_sources/witness_data_source.dart' as _i26;
+    as _i34;
+import '../../identity/data/data_sources/wallet_data_source.dart' as _i25;
 import '../../identity/data/mappers/auth_request_mapper.dart' as _i3;
 import '../../identity/data/mappers/auth_response_mapper.dart' as _i4;
 import '../../identity/data/mappers/hex_mapper.dart' as _i12;
 import '../../identity/data/mappers/identity_dto_mapper.dart' as _i15;
-import '../../identity/data/mappers/private_key_mapper.dart' as _i18;
-import '../../identity/data/repositories/identity_repository_impl.dart' as _i36;
-import '../../identity/domain/repositories/identity_repository.dart' as _i39;
-import '../../identity/domain/use_cases/authenticate_use_case.dart' as _i42;
-import '../../identity/domain/use_cases/create_identity_use_case.dart' as _i43;
-import '../../identity/domain/use_cases/get_auth_token_use_case.dart' as _i44;
+import '../../identity/data/mappers/private_key_mapper.dart' as _i19;
+import '../../identity/data/repositories/identity_repository_impl.dart' as _i38;
+import '../../identity/domain/repositories/identity_repository.dart' as _i41;
+import '../../identity/domain/use_cases/authenticate_use_case.dart' as _i44;
+import '../../identity/domain/use_cases/create_identity_use_case.dart' as _i45;
+import '../../identity/domain/use_cases/get_auth_token_use_case.dart' as _i46;
 import '../../identity/domain/use_cases/get_current_identifier_use_case.dart'
-    as _i45;
-import '../../identity/domain/use_cases/get_identity_use_case.dart' as _i46;
+    as _i47;
+import '../../identity/domain/use_cases/get_identity_use_case.dart' as _i48;
 import '../../identity/domain/use_cases/remove_current_identity_use_case.dart'
-    as _i40;
-import '../../identity/domain/use_cases/sign_message_use_case.dart' as _i41;
+    as _i42;
+import '../../identity/domain/use_cases/sign_message_use_case.dart' as _i43;
 import '../../identity/libs/bjj/bjj.dart' as _i5;
 import '../../identity/libs/iden3core/iden3core.dart' as _i14;
-import '../../proof_generation/libs/prover/prover.dart' as _i20;
-import '../../proof_generation/libs/witnesscalc/auth/witness_auth.dart' as _i25;
-import '../../proof_generation/libs/witnesscalc/mtp/witness_mtp.dart' as _i27;
-import '../../proof_generation/libs/witnesscalc/sig/witness_sig.dart' as _i28;
-import '../credential_wallet.dart' as _i49;
-import '../identity_wallet.dart' as _i47;
-import 'injector.dart' as _i50; // ignore_for_file: unnecessary_lambdas
+import '../../proof_generation/data/data_sources/atomic_query_inputs_data_source.dart'
+    as _i30;
+import '../../proof_generation/data/data_sources/local_files_data_source.dart'
+    as _i18;
+import '../../proof_generation/data/data_sources/proof_scope_data_source.dart'
+    as _i20;
+import '../../proof_generation/data/data_sources/prover_lib_data_source.dart'
+    as _i22;
+import '../../proof_generation/data/data_sources/witness_data_source.dart'
+    as _i27;
+import '../../proof_generation/libs/prover/prover.dart' as _i21;
+import '../../proof_generation/libs/witnesscalc/auth/witness_auth.dart' as _i26;
+import '../../proof_generation/libs/witnesscalc/mtp/witness_mtp.dart' as _i28;
+import '../../proof_generation/libs/witnesscalc/sig/witness_sig.dart' as _i29;
+import '../credential_wallet.dart' as _i51;
+import '../identity_wallet.dart' as _i49;
+import 'injector.dart' as _i52; // ignore_for_file: unnecessary_lambdas
 
 // ignore_for_file: lines_longer_than_80_chars
 /// initializes the registration of provided dependencies inside of [GetIt]
@@ -89,140 +96,145 @@ _i1.GetIt $initSDKGetIt(_i1.GetIt get,
   gh.factory<_i14.Iden3CoreLib>(() => _i14.Iden3CoreLib());
   gh.factory<_i15.IdentityDTOMapper>(() => _i15.IdentityDTOMapper());
   gh.factory<_i16.JWZIsolatesWrapper>(() => _i16.JWZIsolatesWrapper());
-  gh.factory<_i17.LibIdentityWrapper>(
-      () => _i17.LibIdentityWrapper(get<_i14.Iden3CoreLib>()));
-  gh.factory<_i18.PrivateKeyMapper>(() => _i18.PrivateKeyMapper());
-  gh.factory<_i19.ProofScopeDataSource>(() => _i19.ProofScopeDataSource());
-  gh.factory<_i20.ProverLib>(() => _i20.ProverLib());
-  gh.factory<_i21.ProverLibWrapper>(() => _i21.ProverLibWrapper());
-  gh.factory<_i22.RemoteClaimDataSource>(
-      () => _i22.RemoteClaimDataSource(get<_i7.Client>()));
-  gh.factory<_i23.RemoteIdentityDataSource>(
-      () => _i23.RemoteIdentityDataSource(get<_i7.Client>()));
-  gh.factory<_i9.StoreRef<String, Map<String, Object?>>>(
-      () => databaseModule.identityStore,
-      instanceName: 'identityStore');
+  gh.factory<_i17.LibIdentityDataSource>(
+      () => _i17.LibIdentityDataSource(get<_i14.Iden3CoreLib>()));
+  gh.factory<_i18.LocalFilesDataSource>(() => _i18.LocalFilesDataSource());
+  gh.factory<_i19.PrivateKeyMapper>(() => _i19.PrivateKeyMapper());
+  gh.factory<_i20.ProofScopeDataSource>(() => _i20.ProofScopeDataSource());
+  gh.factory<_i21.ProverLib>(() => _i21.ProverLib());
+  gh.factory<_i22.ProverLibWrapper>(() => _i22.ProverLibWrapper());
+  gh.factory<_i23.RemoteClaimDataSource>(
+      () => _i23.RemoteClaimDataSource(get<_i7.Client>()));
+  gh.factory<_i24.RemoteIdentityDataSource>(
+      () => _i24.RemoteIdentityDataSource(get<_i7.Client>()));
   gh.factory<_i9.StoreRef<String, Map<String, Object?>>>(
       () => databaseModule.claimStore,
       instanceName: 'claimStore');
   gh.factory<_i9.StoreRef<String, dynamic>>(() => databaseModule.keyValueStore,
       instanceName: 'keyValueStore');
-  gh.factory<_i24.WalletLibWrapper>(() => _i24.WalletLibWrapper());
-  gh.factory<_i25.WitnessAuthLib>(() => _i25.WitnessAuthLib());
-  gh.factory<_i26.WitnessIsolatesWrapper>(() => _i26.WitnessIsolatesWrapper());
-  gh.factory<_i27.WitnessMtpLib>(() => _i27.WitnessMtpLib());
-  gh.factory<_i28.WitnessSigLib>(() => _i28.WitnessSigLib());
-  gh.factory<_i29.ClaimMapper>(
-      () => _i29.ClaimMapper(get<_i6.ClaimStateMapper>()));
-  gh.factory<_i30.ClaimStoreRefWrapper>(() => _i30.ClaimStoreRefWrapper(
+  gh.factory<_i9.StoreRef<String, Map<String, Object?>>>(
+      () => databaseModule.identityStore,
+      instanceName: 'identityStore');
+  gh.factory<_i25.WalletLibWrapper>(() => _i25.WalletLibWrapper());
+  gh.factory<_i26.WitnessAuthLib>(() => _i26.WitnessAuthLib());
+  gh.factory<_i27.WitnessIsolatesWrapper>(() => _i27.WitnessIsolatesWrapper());
+  gh.factory<_i28.WitnessMtpLib>(() => _i28.WitnessMtpLib());
+  gh.factory<_i29.WitnessSigLib>(() => _i29.WitnessSigLib());
+  gh.factory<_i30.AtomicQueryInputsWrapper>(
+      () => _i30.AtomicQueryInputsWrapper(get<_i14.Iden3CoreLib>()));
+  gh.factory<_i31.ClaimMapper>(
+      () => _i31.ClaimMapper(get<_i6.ClaimStateMapper>()));
+  gh.factory<_i32.ClaimStoreRefWrapper>(() => _i32.ClaimStoreRefWrapper(
       get<_i9.StoreRef<String, Map<String, Object?>>>(
           instanceName: 'claimStore')));
-  gh.factory<_i31.IdentityStoreRefWrapper>(() => _i31.IdentityStoreRefWrapper(
+  gh.factory<_i33.IdentityStoreRefWrapper>(() => _i33.IdentityStoreRefWrapper(
       get<_i9.StoreRef<String, Map<String, Object?>>>(
           instanceName: 'identityStore')));
-  gh.factory<_i32.KeyValueStoreRefWrapper>(() => _i32.KeyValueStoreRefWrapper(
+  gh.factory<_i34.KeyValueStoreRefWrapper>(() => _i34.KeyValueStoreRefWrapper(
       get<_i9.StoreRef<String, dynamic>>(instanceName: 'keyValueStore')));
-  gh.factory<_i17.LibIdentityDataSource>(() => _i17.LibIdentityDataSource(
-      get<_i14.Iden3CoreLib>(), get<_i17.LibIdentityWrapper>()));
-  gh.factory<_i21.ProverLibDataSource>(
-      () => _i21.ProverLibDataSource(get<_i21.ProverLibWrapper>()));
-  gh.factoryAsync<_i30.StorageClaimDataSource>(() async =>
-      _i30.StorageClaimDataSource(await get.getAsync<_i9.Database>(),
-          get<_i30.ClaimStoreRefWrapper>()));
-  gh.factoryAsync<_i32.StorageKeyValueDataSource>(() async =>
-      _i32.StorageKeyValueDataSource(await get.getAsync<_i9.Database>(),
-          get<_i32.KeyValueStoreRefWrapper>()));
-  gh.factory<_i24.WalletDataSource>(
-      () => _i24.WalletDataSource(get<_i24.WalletLibWrapper>()));
-  gh.factory<_i26.WitnessDataSource>(
-      () => _i26.WitnessDataSource(get<_i26.WitnessIsolatesWrapper>()));
-  gh.factoryAsync<_i33.CredentialRepositoryImpl>(() async =>
-      _i33.CredentialRepositoryImpl(
-          get<_i22.RemoteClaimDataSource>(),
-          await get.getAsync<_i30.StorageClaimDataSource>(),
+  gh.factory<_i22.ProverLibDataSource>(
+      () => _i22.ProverLibDataSource(get<_i22.ProverLibWrapper>()));
+  gh.factoryAsync<_i32.StorageClaimDataSource>(() async =>
+      _i32.StorageClaimDataSource(await get.getAsync<_i9.Database>(),
+          get<_i32.ClaimStoreRefWrapper>()));
+  gh.factoryAsync<_i34.StorageKeyValueDataSource>(() async =>
+      _i34.StorageKeyValueDataSource(await get.getAsync<_i9.Database>(),
+          get<_i34.KeyValueStoreRefWrapper>()));
+  gh.factory<_i25.WalletDataSource>(
+      () => _i25.WalletDataSource(get<_i25.WalletLibWrapper>()));
+  gh.factory<_i27.WitnessDataSource>(
+      () => _i27.WitnessDataSource(get<_i27.WitnessIsolatesWrapper>()));
+  gh.factory<_i30.AtomicQueryInputsDataSource>(() =>
+      _i30.AtomicQueryInputsDataSource(get<_i30.AtomicQueryInputsWrapper>()));
+  gh.factoryAsync<_i35.CredentialRepositoryImpl>(() async =>
+      _i35.CredentialRepositoryImpl(
+          get<_i23.RemoteClaimDataSource>(),
+          await get.getAsync<_i32.StorageClaimDataSource>(),
           get<_i8.CredentialRequestMapper>(),
-          get<_i29.ClaimMapper>(),
+          get<_i31.ClaimMapper>(),
           get<_i11.FiltersMapper>(),
           get<_i13.IdFilterMapper>()));
   gh.factory<_i16.JWZDataSource>(() => _i16.JWZDataSource(
       get<_i5.BabyjubjubLib>(),
-      get<_i24.WalletDataSource>(),
+      get<_i25.WalletDataSource>(),
       get<_i16.JWZIsolatesWrapper>()));
-  gh.factoryAsync<_i31.StorageIdentityDataSource>(() async =>
-      _i31.StorageIdentityDataSource(
+  gh.factoryAsync<_i33.StorageIdentityDataSource>(() async =>
+      _i33.StorageIdentityDataSource(
           await get.getAsync<_i9.Database>(),
-          get<_i31.IdentityStoreRefWrapper>(),
-          await get.getAsync<_i32.StorageKeyValueDataSource>()));
-  gh.factoryAsync<_i34.CredentialRepository>(() async =>
+          get<_i33.IdentityStoreRefWrapper>(),
+          await get.getAsync<_i34.StorageKeyValueDataSource>()));
+  gh.factoryAsync<_i36.CredentialRepository>(() async =>
       repositoriesModule.credentialRepository(
-          await get.getAsync<_i33.CredentialRepositoryImpl>()));
-  gh.factoryAsync<_i35.GetClaimsUseCase>(() async =>
-      _i35.GetClaimsUseCase(await get.getAsync<_i34.CredentialRepository>()));
-  gh.factoryAsync<_i36.IdentityRepositoryImpl>(() async =>
-      _i36.IdentityRepositoryImpl(
-          get<_i24.WalletDataSource>(),
+          await get.getAsync<_i35.CredentialRepositoryImpl>()));
+  gh.factoryAsync<_i37.GetClaimsUseCase>(() async =>
+      _i37.GetClaimsUseCase(await get.getAsync<_i36.CredentialRepository>()));
+  gh.factoryAsync<_i38.IdentityRepositoryImpl>(() async =>
+      _i38.IdentityRepositoryImpl(
+          get<_i25.WalletDataSource>(),
           get<_i17.LibIdentityDataSource>(),
-          await get.getAsync<_i31.StorageIdentityDataSource>(),
-          await get.getAsync<_i32.StorageKeyValueDataSource>(),
+          await get.getAsync<_i33.StorageIdentityDataSource>(),
+          await get.getAsync<_i34.StorageKeyValueDataSource>(),
           get<_i16.JWZDataSource>(),
           get<_i12.HexMapper>(),
-          get<_i18.PrivateKeyMapper>(),
+          get<_i19.PrivateKeyMapper>(),
           get<_i15.IdentityDTOMapper>(),
-          get<_i23.RemoteIdentityDataSource>(),
+          get<_i24.RemoteIdentityDataSource>(),
           get<_i3.AuthRequestMapper>(),
-          get<_i19.ProofScopeDataSource>(),
-          await get.getAsync<_i30.StorageClaimDataSource>(),
-          get<_i29.ClaimMapper>(),
+          get<_i20.ProofScopeDataSource>(),
+          await get.getAsync<_i32.StorageClaimDataSource>(),
+          get<_i31.ClaimMapper>(),
           get<_i11.FiltersMapper>(),
-          get<_i26.WitnessDataSource>(),
-          get<_i21.ProverLibDataSource>(),
-          get<_i4.AuthResponseMapper>()));
-  gh.factoryAsync<_i37.RemoveClaimsUseCase>(() async =>
-      _i37.RemoveClaimsUseCase(
-          await get.getAsync<_i34.CredentialRepository>()));
-  gh.factoryAsync<_i38.UpdateClaimUseCase>(() async =>
-      _i38.UpdateClaimUseCase(await get.getAsync<_i34.CredentialRepository>()));
-  gh.factoryAsync<_i39.IdentityRepository>(() async => repositoriesModule
-      .identityRepository(await get.getAsync<_i36.IdentityRepositoryImpl>()));
-  gh.factoryAsync<_i40.RemoveCurrentIdentityUseCase>(() async =>
-      _i40.RemoveCurrentIdentityUseCase(
-          await get.getAsync<_i39.IdentityRepository>()));
-  gh.factoryAsync<_i41.SignMessageUseCase>(() async =>
-      _i41.SignMessageUseCase(await get.getAsync<_i39.IdentityRepository>()));
-  gh.factoryAsync<_i42.AuthenticateUseCase>(() async =>
-      _i42.AuthenticateUseCase(await get.getAsync<_i39.IdentityRepository>()));
-  gh.factoryAsync<_i43.CreateIdentityUseCase>(() async =>
-      _i43.CreateIdentityUseCase(
-          await get.getAsync<_i39.IdentityRepository>()));
-  gh.factoryAsync<_i44.GetAuthTokenUseCase>(() async =>
-      _i44.GetAuthTokenUseCase(await get.getAsync<_i39.IdentityRepository>()));
-  gh.factoryAsync<_i45.GetCurrentIdentifierUseCase>(() async =>
-      _i45.GetCurrentIdentifierUseCase(
-          await get.getAsync<_i39.IdentityRepository>()));
-  gh.factoryAsync<_i46.GetIdentityUseCase>(() async =>
-      _i46.GetIdentityUseCase(await get.getAsync<_i39.IdentityRepository>()));
-  gh.factoryAsync<_i47.IdentityWallet>(() async => _i47.IdentityWallet(
-      await get.getAsync<_i43.CreateIdentityUseCase>(),
-      await get.getAsync<_i46.GetIdentityUseCase>(),
-      await get.getAsync<_i41.SignMessageUseCase>(),
-      await get.getAsync<_i44.GetAuthTokenUseCase>(),
-      await get.getAsync<_i45.GetCurrentIdentifierUseCase>(),
-      await get.getAsync<_i40.RemoveCurrentIdentityUseCase>(),
-      await get.getAsync<_i42.AuthenticateUseCase>()));
-  gh.factoryAsync<_i48.FetchAndSaveClaimsUseCase>(() async =>
-      _i48.FetchAndSaveClaimsUseCase(
-          await get.getAsync<_i44.GetAuthTokenUseCase>(),
-          await get.getAsync<_i34.CredentialRepository>()));
-  gh.factoryAsync<_i49.CredentialWallet>(() async => _i49.CredentialWallet(
-      await get.getAsync<_i48.FetchAndSaveClaimsUseCase>(),
-      await get.getAsync<_i35.GetClaimsUseCase>(),
-      await get.getAsync<_i37.RemoveClaimsUseCase>(),
-      await get.getAsync<_i38.UpdateClaimUseCase>()));
+          get<_i27.WitnessDataSource>(),
+          get<_i22.ProverLibDataSource>(),
+          get<_i4.AuthResponseMapper>(),
+          get<_i30.AtomicQueryInputsDataSource>(),
+          get<_i18.LocalFilesDataSource>()));
+  gh.factoryAsync<_i39.RemoveClaimsUseCase>(() async =>
+      _i39.RemoveClaimsUseCase(
+          await get.getAsync<_i36.CredentialRepository>()));
+  gh.factoryAsync<_i40.UpdateClaimUseCase>(() async =>
+      _i40.UpdateClaimUseCase(await get.getAsync<_i36.CredentialRepository>()));
+  gh.factoryAsync<_i41.IdentityRepository>(() async => repositoriesModule
+      .identityRepository(await get.getAsync<_i38.IdentityRepositoryImpl>()));
+  gh.factoryAsync<_i42.RemoveCurrentIdentityUseCase>(() async =>
+      _i42.RemoveCurrentIdentityUseCase(
+          await get.getAsync<_i41.IdentityRepository>()));
+  gh.factoryAsync<_i43.SignMessageUseCase>(() async =>
+      _i43.SignMessageUseCase(await get.getAsync<_i41.IdentityRepository>()));
+  gh.factoryAsync<_i44.AuthenticateUseCase>(() async =>
+      _i44.AuthenticateUseCase(await get.getAsync<_i41.IdentityRepository>()));
+  gh.factoryAsync<_i45.CreateIdentityUseCase>(() async =>
+      _i45.CreateIdentityUseCase(
+          await get.getAsync<_i41.IdentityRepository>()));
+  gh.factoryAsync<_i46.GetAuthTokenUseCase>(() async =>
+      _i46.GetAuthTokenUseCase(await get.getAsync<_i41.IdentityRepository>()));
+  gh.factoryAsync<_i47.GetCurrentIdentifierUseCase>(() async =>
+      _i47.GetCurrentIdentifierUseCase(
+          await get.getAsync<_i41.IdentityRepository>()));
+  gh.factoryAsync<_i48.GetIdentityUseCase>(() async =>
+      _i48.GetIdentityUseCase(await get.getAsync<_i41.IdentityRepository>()));
+  gh.factoryAsync<_i49.IdentityWallet>(() async => _i49.IdentityWallet(
+      await get.getAsync<_i45.CreateIdentityUseCase>(),
+      await get.getAsync<_i48.GetIdentityUseCase>(),
+      await get.getAsync<_i43.SignMessageUseCase>(),
+      await get.getAsync<_i46.GetAuthTokenUseCase>(),
+      await get.getAsync<_i47.GetCurrentIdentifierUseCase>(),
+      await get.getAsync<_i42.RemoveCurrentIdentityUseCase>(),
+      await get.getAsync<_i44.AuthenticateUseCase>()));
+  gh.factoryAsync<_i50.FetchAndSaveClaimsUseCase>(() async =>
+      _i50.FetchAndSaveClaimsUseCase(
+          await get.getAsync<_i46.GetAuthTokenUseCase>(),
+          await get.getAsync<_i36.CredentialRepository>()));
+  gh.factoryAsync<_i51.CredentialWallet>(() async => _i51.CredentialWallet(
+      await get.getAsync<_i50.FetchAndSaveClaimsUseCase>(),
+      await get.getAsync<_i37.GetClaimsUseCase>(),
+      await get.getAsync<_i39.RemoveClaimsUseCase>(),
+      await get.getAsync<_i40.UpdateClaimUseCase>()));
   return get;
 }
 
-class _$NetworkModule extends _i50.NetworkModule {}
+class _$NetworkModule extends _i52.NetworkModule {}
 
-class _$DatabaseModule extends _i50.DatabaseModule {}
+class _$DatabaseModule extends _i52.DatabaseModule {}
 
-class _$RepositoriesModule extends _i50.RepositoriesModule {}
+class _$RepositoriesModule extends _i52.RepositoriesModule {}
